# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for TechGuild Platform.",
        "title": "TechGuild Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "TechGuild Team",
            "email": "support@techguild.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "https://techguild-backend.onrender.com",
    "basePath": "/",
    "paths": {
        "/admin/verification/queue": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the list of pending verification requests for administrators.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Verification"
                ],
                "summary": "Get verification queue",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.VerificationQueueItem"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/verification/{id}/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Approves a verification request.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Verification"
                ],
                "summary": "Approve verification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Verification Record ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AdminApproveResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/verification/{id}/reject": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Rejects a verification request with a reason.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Verification"
                ],
                "summary": "Reject verification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Verification Record ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Rejection reason",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AdminRejectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AdminRejectResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/verification/business/submit": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Submits a business verification request with the required business documents.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Verification"
                ],
                "summary": "Submit business verification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Business Name",
                        "name": "business_name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "GST Number",
                        "name": "gst_number",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "PAN Number",
                        "name": "pan_number",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "GST Certificate",
                        "name": "gst_certificate",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "PAN Card",
                        "name": "pan_card",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Authorized Representative ID",
                        "name": "authorized_representative_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.BusinessVerificationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/verification/identity/submit": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Submits an individual's identity verification request with government ID and selfie.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Verification"
                ],
                "summary": "Submit identity verification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Full Name",
                        "name": "full_name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Government ID Type",
                        "name": "govt_id_type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Government ID Number",
                        "name": "govt_id_number",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Additional verification data",
                        "name": "profile_data",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Government ID Document",
                        "name": "govt_id_document",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Selfie",
                        "name": "selfie",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.IdentityVerificationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/verification/resubmit/{record_id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Resubmits a rejected verification request.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Verification"
                ],
                "summary": "Resubmit verification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Verification Record ID",
                        "name": "record_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Additional comments",
                        "name": "reason",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Updated Government ID Document",
                        "name": "govt_id_document",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ResubmitVerificationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.AdminApproveResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.AdminRejectRequest": {
            "type": "object",
            "required": [
                "reason"
            ],
            "properties": {
                "reason": {
                    "type": "string"
                }
            }
        },
        "dto.AdminRejectResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.BusinessVerificationResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "verification_record_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.IdentityVerificationResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "verification_record_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.ResubmitVerificationResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "verification_record_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.VerificationQueueItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}